Restructure project into modular layout
