+++ README.md
@@ -1,22 +1,24 @@
 fs25-discord-bot
+=================
 
-Бот отображает технику, требующую обслуживания. Техника делится на три
-категории:
+Асинхронный Discord-бот для отображения техники, требующей обслуживания.
 
-* **Повреждённая техника (низкое топливо + повреждение)** — техника с
-  повреждением более 50% или топливом ниже 40%.
-* **Сильно загрязнённая техника** — техника с загрязнением более 50%. Если
-  повреждения меньше 5%, этот параметр не отображается.
-* **Остальная техника** — техника с повреждением или загрязнением более 5%.
-  Если один из параметров ниже 5%, он не выводится.
+Техника делится на три категории:
 
-Загрузчик FTP-файлов
+* **Повреждённая техника (низкое топливо + повреждение)** — повреждение более 50% или топливо ниже 40%.
+* **Сильно загрязнённая техника** — загрязнение более 50%. Если повреждения меньше 5%, этот параметр не отображается.
+* **Остальная техника** — повреждение или загрязнение более 5%. Если один из параметров ниже 5%, он не выводится.
 
-Скрипт ftp_upload.py позволяет загружать произвольные файлы на FTP-сервер.
-Параметры подключения можно передавать через аргументы командной строки
-или переменные окружения FTP_HOST, FTP_PORT, FTP_USER, FTP_PASS.
+Все параметры подключения и пути настраиваются через `config/config.py` или переменные окружения.
 
-Пример запуска:
+Запуск бота:
 
-python ftp_upload.py local.zip /remote/path/remote.zip --host example.com --user user --password pass
+```bash
+python -m bot.main
+```
 
+Загрузка файлов на FTP сервер:
+
+```bash
+python -m ftp.upload local.zip /remote/path/remote.zip
+```
