+++ modules/chats.py
@@ -1,0 +1,5 @@
+"""Placeholder for chat log processing."""
+
+
+def process_chat_logs(data: bytes) -> str:
+    return "Обработка чатов пока не реализована"
