+++ modules/fields.py
@@ -1,0 +1,5 @@
+"""Placeholder for field analysis logic."""
+
+
+def analyze_fields(data: bytes) -> str:
+    return "Анализ полей пока не реализован"
