+++ requirements.txt
@@ -1,1 +1,2 @@
 py-cord==2.4.1
+aioftp==0.22.0
