--- runtime.txt
@@ -1,1 +1,0 @@
-python-3.11.8
